Enum ferrous_big_decimal::digits::Digits

pub enum Digits {
    Zero,
    One,
    Two,
    Three,
    Four,
    Five,
    Six,
    Seven,
    Eight,
    Nine,
}

Variants

Zero

One

Two

Three

Four

Five

Six

Seven

Eight

Nine

Implementations

impl Digits

pub const fn addition(&self, rhs: Digits) -> (Digits, Digits)

Adds rhs to lhs with any carry over ammount returned.

pub const fn fused_addition(&self, rhs: Digits, carry: Digits) -> (Digits, Digits)

Adds rhs and carry to self with any borrowed amount returned.

pub const fn complement(x: Digits) -> Digits

Returns the complement of a Digit.

Used to speed up subtraction.

pub const fn subtract(&self, rhs: Digits) -> (Digits, Digits)

Subtracts rhs from self with any borrowed ammount returned.

pub const fn fused_subtraction(
&self,
rhs: Digits,
carry: Digits
) -> (Digits, Digits)

pub const fn multiply(&self, rhs: Digits) -> (Digits, Digits)

Multiplies self by rhs with any carry over amount returned.

pub const fn divide(&self, rhs: Digits) -> Result<(Digits, Digits), MathErrors>

Divides self by rhs with any remainder returned.

MathErrors::DivisionByZero is returned if self or rhs are 0.

pub const fn to_usize(&self) -> usize

pub const fn to_char(self) -> char